svg/struct/UnknownElement/interface.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="100%" height="100%" version="1.1" viewBox="0 0 400 400"><title>Interface of unknown element</title><h:link href="https://svgwg.org/svg2-draft/struct.html#UnknownElement" rel="help"/><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><foobar id="target"/><h:script><![CDATA[
test(function() {
    var e = document.getElementById("target");
    assert_equals(e.__proto__, SVGUnknownElement.prototype);
});
]]></h:script></svg>
